Shop improvements + added review system
Please note the view system is all inside a class. The code needs to be
moved outside of the /assets/ directory and placed in /classes/
directory and the viewing code must be moved into /views/v-rating.php
later on.

The .htaccess file needs to allow clean urls on this page along with
this system been locked to only users who have admin access on the site.